move code into /src, transpile with babel into /lib, point to /lib files in package.json, change namespace to @loopmode/dis-gui
