+++ .gitignore
@@ -36,3 +36,5 @@ jspm_packages
 
 # Optional REPL history
 .node_repl_history
+
+lib
