+++ package.json
@@ -1,11 +1,12 @@
 {
-  "name": "dis-gui",
+  "name": "@loopmode/dis-gui",
   "version": "2.0.0",
   "description": "An extensible, styleable, & React-based controller library inspired by the venerable dat-gui.",
-  "main": "index.js",
+  "main": "lib/index.js",
   "scripts": {
     "start": "budo example/main.js:bundle.js -d example --live -- -t [ babelify --presets [ es2015 react ] ]",
-    "build": "browserify example/main.js -o gh-pages/bundle.js -t [ babelify --presets [ es2015 react ] ]"
+    "build:example": "browserify example/main.js -o gh-pages/bundle.js -t [ babelify --presets [ es2015 react ] ]",
+    "build": "babel src --out-dir lib"
   },
   "license": "Unlicense",
   "dependencies": {
@@ -19,29 +20,17 @@
     "react-dom": "^15.3.0"
   },
   "devDependencies": {
-    "babel-preset-es2015": "^6.13.2",
-    "babel-preset-react": "^6.11.1",
     "babelify": "^7.3.0",
     "browserify": "^13.1.0",
     "budo": "^9.0.0",
-    "gl-mat4": "^1.1.4",
-    "react": "^15.3.0",
-    "react-addons-update": "^15.3.1",
-    "react-dom": "^15.3.0",
-    "regl": "^1.1.1"
-  },
-  "browserify": {
-    "transform": [
-      [
-        "babelify",
-        {
-          "presets": [
-            "es2015",
-            "react"
-          ]
-        }
-      ]
-    ]
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.25.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-register": "^6.24.1",
+    "esdoc": "^0.5.2",
+    "regl": "^1.1.1",
+    "gl-mat4": "^1.1.4"
   },
   "keywords": [
     "dat-gui",
