+++ src/components/components/color-range.jsx
