+++ src/components/gradient/stop.jsx
