+++ src/components/text.jsx
